[arcane,CSV,RNG] Add fix namespace in the two services -- Edit encoding of all edited files
