+++ arcane/src/arcane/IRandomNumberGenerator.h
@@ -14,11 +14,20 @@
 #ifndef ARCANE_IRANDOMNUMBERGENERATOR_H
 #define ARCANE_IRANDOMNUMBERGENERATOR_H
 
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 #include "arcane/utils/Array.h"
 #include "arcane/utils/UtilsTypes.h"
 
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 namespace Arcane
 {
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
 
 /**
  * @brief Classe représentant une graine.
@@ -296,6 +305,13 @@ class ARCANE_CORE_EXPORT IRandomNumberGenerator
    */
   virtual Real generateRandomNumber(RandomNumberGeneratorSeed* seed, Integer leap = 0) = 0;
 };
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 } // End namespace Arcane
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
 
 #endif
