+++ arcane/src/arcane/ISimpleTableOutput.h
@@ -14,12 +14,21 @@
 #ifndef ARCANE_ISIMPLEOUTPUT_H
 #define ARCANE_ISIMPLEOUTPUT_H
 
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 #include <arcane/ItemTypes.h>
 
 #include <optional>
 
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 namespace Arcane
 {
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
 
 /**
  * @ingroup StandardService
@@ -656,6 +665,15 @@ public:
   virtual bool writeFile(String path, Integer only_proc = -1) = 0;
 };
 
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 } // End namespace Arcane
 
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 #endif
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
