+++ arcane/src/arcane/std/PDESRandomNumberGeneratorService.cc
@@ -19,7 +19,11 @@
 
 #include "arcane/std/PDESRandomNumberGeneratorService.h"
 
-using namespace Arcane;
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+namespace Arcane
+{
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -218,3 +222,11 @@ _ran4(Int64* seed, Integer leap)
   volatile Real fin = 5.4210108624275222e-20 * (Real)_reconstructUInt64(lword, irword);
   return fin;
 }
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+} // End namespace Arcane
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
