+++ arcane/src/arcane/tests/SimpleTableOutputUnitTest.axl
@@ -1,5 +1,5 @@
-﻿<?xml version="1.0" ?>
-<service name="SimpleTableOutputUnitTest" version="1.0" type="caseoption" parent-name="Arcane::BasicUnitTest"> 
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<service name="SimpleTableOutputUnitTest" version="1.0" type="caseoption" parent-name="Arcane::BasicUnitTest" namespace-name="ArcaneTest"> 
   <interface name="Arcane::IUnitTest" inherited='false'/>
 
   <description>Descripteur du service SimpleTableOutputUnitTest</description>
