+++ arcane/src/arcane/tests/SimpleTableOutputUnitTest.cc
@@ -12,7 +12,15 @@
 /*---------------------------------------------------------------------------*/
 
 #include "arcane/tests/SimpleTableOutputUnitTest.h"
-#include <iostream>
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+namespace ArcaneTest
+{
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
 
 template<class T>
 void SimpleTableOutputUnitTest::
@@ -23,6 +31,9 @@ ASSERT_EQUAL_ARRAY(UniqueArray<T> expected, UniqueArray<T> actual)
     ASSERT_EQUAL(expected[i], actual[i]);
   }
 }
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
 
 void SimpleTableOutputUnitTest::
 setUpForClass()
@@ -1048,6 +1059,11 @@ tearDownForClass()
 {
 
 }
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+} // End namespace ArcaneTest
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
