+++ arcane/src/arcane/tests/SimpleTableOutputUnitTest.h
@@ -17,10 +17,21 @@
 #include "arcane/BasicUnitTest.h"
 #include "arcane/ISimpleTableOutput.h"
 
-
 #include "arcane/tests/SimpleTableOutputUnitTest_axl.h"
 
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+namespace ArcaneTest
+{
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 using namespace Arcane;
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
 
 class SimpleTableOutputUnitTest
   : public ArcaneSimpleTableOutputUnitTestObject
@@ -115,6 +126,11 @@ private:
 /*---------------------------------------------------------------------------*/
 
 ARCANE_REGISTER_SERVICE_SIMPLETABLEOUTPUTUNITTEST(SimpleTableOutputUnitTest, SimpleTableOutputUnitTest);
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+} // End namespace ArcaneTest
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
